feat(arcteryx): add price
