+++ lib/v2/arcteryx/new-arrivals.js
@@ -1,21 +1,8 @@
 const got = require('@/utils/got');
 const { art } = require('@/utils/render');
 const path = require('path');
+const { generateRssData } = require('./utils');
 
-const attributeSet = new Set(['name', 'image', 'short_description', 'slug']);
-function generateRssData(item, index, arr) {
-    const attributes = item.attribute;
-    const data = {};
-
-    attributes.forEach((attribute) => {
-        const key = attribute.name;
-        const value = attribute.value[0].value;
-        if (attributeSet.has(key)) {
-            data[key] = value;
-        }
-    });
-    arr[index] = data;
-}
 module.exports = async (ctx) => {
     const { country, gender } = ctx.params;
     const host = `https://arcteryx.com/${country}/en/`;
@@ -32,7 +19,9 @@ module.exports = async (ctx) => {
         },
     });
     const items = response.data.universes.universe[1]['items-section'].items.item;
-    items.forEach(generateRssData);
+    items.forEach((item, index, arr) => {
+        generateRssData(item, index, arr, country);
+    });
 
     ctx.state.data = {
         title: `Arcteryx - New Arrivals(${country.toUpperCase()}) - ${gender.toUpperCase()}`,
