+++ lib/v2/arcteryx/templates/product-description.art
@@ -1,6 +1,12 @@
 <div>
+    {{if item.short_description}}
+    {{item.short_description}}<br>
+    {{/if}}
+    {{if item.original_price}}
+    Original Price: {{item.original_price}}<br>
+    {{/if}}
+    {{if item.price}}
+    Current Price: {{item.price}}<br>
+    {{/if}}
     <img src={{item.image}}>
-    <br><br>
-    {{item.short_description}}
-    <br><br>
 </div>
